+++ src/data_type/mod.rs
@@ -64,6 +64,7 @@ use crate::{
     namer,
     types::{And, Or},
     visitor::{self, Acceptor},
+    hierarchy::{Hierarchy, Path}
 };
 use injection::{Base, InjectInto, Injection};
 use intervals::{Bound, Intervals};
@@ -859,7 +860,6 @@ impl Struct {
     pub fn field_from_index(&self, index: usize) -> &(String, Rc<DataType>) {
         &self.fields[index]
     }
-
     /// Build the type of a DataFrame (column based data)
     pub fn from_schema_size(schema: Struct, size: &Integer) -> Struct {
         schema
@@ -867,6 +867,16 @@ impl Struct {
             .into_iter()
             .map(|(s, t)| (s, Rc::new(List::new(t, size.clone()).into())))
             .collect()
+    }
+    // TODO This could be implemented with a visitor (it would not fail on cyclic cases)
+    /// Produce a Hierarchy of subtypes to access them in a smart way (unambiguous prefix can be omited)
+    pub fn hierarchy(&self) -> Hierarchy<&DataType> {
+        self.iter()
+        .fold(
+            self.iter().map(|(s, d)| (vec![s.clone()], d.as_ref())).collect(),
+            |h, (s, d)|
+            h.chain(d.hierarchy().prepend(&[s.clone()]))
+        )
     }
 }
 
@@ -1068,12 +1078,11 @@ impl InjectInto<DataType> for Struct {
 }
 
 // Index Structs
+impl<P: Path> Index<P> for Struct {
+    type Output = DataType;
 
-impl Index<&str> for Struct {
-    type Output = Rc<DataType>;
-
-    fn index(&self, index: &str) -> &Self::Output {
-        &self.field(index).unwrap().1
+    fn index(&self, index: P) -> &Self::Output {
+        self.hierarchy()[index]
     }
 }
 
@@ -1143,6 +1152,16 @@ impl Union {
     /// Access a field by index
     pub fn field_from_index(&self, index: usize) -> &(String, Rc<DataType>) {
         &self.fields[index]
+    }
+    // TODO This could be implemented with a visitor (it would not fail on cyclic cases)
+    /// Produce a Hierarchy of subtypes to access them in a smart way (unambiguous prefix can be omited)
+    pub fn hierarchy(&self) -> Hierarchy<&DataType> {
+        self.iter()
+        .fold(
+            self.iter().map(|(s, d)| (vec![s.clone()], d.as_ref())).collect(),
+            |h, (s, d)|
+            h.chain(d.hierarchy().prepend(&[s.clone()]))
+        )
     }
 }
 
@@ -1348,12 +1367,11 @@ impl InjectInto<DataType> for Union {
 }
 
 // Index Unions
+impl<P: Path> Index<P> for Union {
+    type Output = DataType;
 
-impl Index<&str> for Union {
-    type Output = Rc<DataType>;
-
-    fn index(&self, index: &str) -> &Self::Output {
-        &self.field(index).unwrap().1
+    fn index(&self, index: P) -> &Self::Output {
+        self.hierarchy()[index]
     }
 }
 
@@ -2306,6 +2324,17 @@ impl DataType {
             _ => Err(Error::other("No common variant")),
         }
     }
+    // TODO This could be implemented with a visitor (it would not fail on cyclic cases)
+    /// Produce a Hierarchy of subtypes to access them in a smart way (unambiguous prefix can be omited)
+    fn hierarchy(&self) -> Hierarchy<&DataType> {
+        for_all_variants!(
+            self,
+            x,
+            x.hierarchy(),
+            [Struct, Union],
+            Hierarchy::from([(Vec::<&str>::new(), self)])
+        )
+    }
 }
 
 impl Variant for DataType {
@@ -2701,6 +2730,14 @@ impl DataType {
 
     pub fn function(domain: DataType, co_domain: DataType) -> DataType {
         DataType::from(Function::from((domain, co_domain)))
+    }
+}
+
+impl<P: Path> Index<P> for DataType {
+    type Output = DataType;
+
+    fn index(&self, index: P) -> &Self::Output {
+        self.hierarchy()[index]
     }
 }
 
@@ -3324,20 +3361,39 @@ mod tests {
     }
 
     #[test]
-    fn test_struct_index() {
-        let a = DataType::unit() & DataType::boolean() & DataType::boolean() & DataType::float();
-        println!("a = {}", &a);
-        let b = DataType::unit()
-            & ("a", DataType::boolean())
-            & DataType::unit()
-            & a
-            & ("c", DataType::boolean())
-            & ("d", DataType::float());
-        let b = Struct::try_from(b).unwrap();
-        println!("b = {b}");
-        println!("b[4] = {}", b[4]);
-        println!("b[c] = {}", b["c"]);
-        assert_eq!(b[4].as_ref(), &(DataType::float()));
+    fn test_index() {
+        let dt = DataType::float();
+        assert_eq!(dt[Vec::<String>::new()], dt);
+        let dt1 = DataType::structured([
+            ("a", DataType::integer()),
+            ("b", DataType::boolean())
+        ]);
+        let dt2 = DataType::structured([
+            ("a", DataType::float()),
+            ("c", DataType::integer())
+        ]);
+        let dt = DataType::Null
+        | ("table1", dt1.clone())
+        | ("table2", dt2.clone());
+        assert_eq!(dt["table1"], dt1);
+        assert_eq!(dt["table2"], dt2);
+        assert_eq!(dt[["table1", "a"]], DataType::integer());
+        assert_eq!(dt[["table2", "a"]], DataType::float());
+        assert_eq!(dt["b"], DataType::boolean());
+        assert_eq!(dt[["c"]], DataType::integer());
+
+        let a = DataType::structured([
+            ("a_0", DataType::integer_min(-10)),
+            ("a_1", DataType::integer()),
+        ]);
+        let b = DataType::structured([
+            ("b_0", DataType::float()),
+            ("b_1", DataType::float_interval(0., 1.)),
+        ]);
+        let x = DataType::structured([("a", a.clone()), ("b", b)]);
+        println!("{}", x);
+        assert_eq!(x[["a"]], a);
+        assert_eq!(x[["a", "a_1"]], DataType::integer());
     }
 
     #[test]
@@ -3810,5 +3866,43 @@ mod tests {
         let left = DataType::integer_interval(3, 7);
         let right = DataType::optional(DataType::integer_interval(2, 5));
         assert!(DataType::into_common_sub_variant(&left, &right).is_err());
+    }
+
+    #[test]
+    fn test_hierarchy(){
+        let dt_float = DataType::float();
+        let dt_int = DataType::integer();
+        let struct_dt = DataType::structured([
+            ("a", DataType::float()),
+            ("b", DataType::integer()),
+        ]);
+        println!("{}", struct_dt.hierarchy());
+        let correct_hierarchy = Hierarchy::from([
+            (vec!["a"], &dt_float),
+            (vec!["b"], &dt_int),
+        ]);
+        assert_eq!(
+            struct_dt.hierarchy(),
+            correct_hierarchy
+        );
+        let struct_dt2 = DataType::structured([
+            ("a", DataType::integer()),
+            ("c", DataType::integer()),
+        ]);
+        let union_dt = DataType::union([
+            ("table1", struct_dt.clone()),
+            ("table2", struct_dt2.clone()),
+        ]);
+        let correct_hierarchy = Hierarchy::from([
+            (vec!["table1"], &struct_dt),
+            (vec!["table2"], &struct_dt2),
+            (vec!["table1", "a"], &dt_float),
+            (vec!["table1", "b"], &dt_int),
+            (vec!["table2", "a"], &dt_int),
+            (vec!["table2", "c"], &dt_int),
+        ]);
+        let h = union_dt.hierarchy();
+        println!("{}", h);
+        assert_eq!(h, correct_hierarchy);
     }
 }